+++ app/controllers/tasks_controller.rb
@@ -2,8 +2,8 @@ class TasksController < ApplicationController
 
   def sitemap
     if current_admin_user && current_admin_user.has_role?(:admin)
-      SitemapQueue.perform_in(3.minutes)
-      flash[:notice] = "Sitemap Scheduled to run in 3 minutes."
+      SitemapQueue.perform_in(2.minutes)
+      flash[:notice] = "Sitemap Scheduled to run in 2 minutes."
     end
 
     respond_to do |format|