+++ config/initializers/scheduler.rb
@@ -1,0 +1,4 @@
+if Rails.env.production?
+  require './lib/scheduler.rb'
+  Publication::FeedScheduler.new.run!
+end