+++ db/migrate/20121010154859_add_type_to_job_stats.rb
@@ -1,0 +1,8 @@
+class AddTypeToJobStats < ActiveRecord::Migration
+  def change
+    remove_column :job_stats, :job_name
+    add_column :job_stats, :type, :string
+    add_index  :job_stats, :type
+    execute "truncate table job_stats"
+  end
+end